+++ sugar_analysis/__init__.py
@@ -11,28 +11,4 @@ Some description.
 #__all__ = [os.path.basename(m).replace('.py', '') for m in glob.glob("sugar/*.py")
 #           if '__init__' not in m]
 
-from . import multilinearfit
-
-from .load_data import load_data_sugar
-
-from .gaussian_process import load_data_bin_gp
-from .gaussian_process import gp_sed 
-
-from .math_toolbox import passage
-from .math_toolbox import passage_error
-from .math_toolbox import svd_inverse
-from .math_toolbox import cholesky_inverse
-from .math_toolbox import biweight_M
-from .math_toolbox import biweight_S
-
-from .cosmology import distance_modulus
-from .extinction import extinctionLaw
-
-from .emfa_analysis import run_emfa_analysis
-from .sed_fitting import multilinearfit
-from .sed_fitting import sugar_fitting
-from .make_sugar_sed import make_sugar
-
-from .build_sugar import build_model
-
-from .hubble_diagram import comp_rms, hubble_salt2
+from .math_toolbox import comp_rms
